Preprocessing data check
Preprocessing now checks each trial for complete data. If a trial is found incomplete, or unprocessable a warning pops up. The warning indicates the trial data location (which contains the participant name as well).

It is up to the user how they wish to handle that participant's data for preprocessing.