batch script for windows/compiled version
This batch script both sets the compatibility mode of the executable to Windows 7, and starts it from the parent directory (it does not cd into the build dir, as I thought would be necessary). The latter seems to mean that the data and experiments folder are created in that parent directory as well, ideal to avoid the compiled mess. Needs testing on other versions of Windows though (only have 10).